Improved ux flow when cancelling lens publish
